requirements.txt
spotipy==2.23.0
google-generativeai>=0.3.2
python-dotenv==1.0.1
flask==3.0.2
requests==2.31.0 